+++ content/keysView.mjs.js
@@ -18,6 +18,10 @@ document.addEventListener("DOMContentLoaded", async () => {
 
 	const keys = await KeyDbProxy.getKeys();
 
+	// Fix `_detectBrowser()` logic of Tabulator in case the `navigator.userAgent` preference is set to an empty string.
+	// @ts-expect-error
+	globalThis.opera = "Firefox";
+
 	/** @type {import("tabulator-tables").Tabulator} */
 	// @ts-expect-error
 	const table = new Tabulator("#keysTable", {