+++ modules/arhVerifier.mjs.js
@@ -114,8 +114,10 @@ function readARHs(arHeaders, account) {
  */
 function arhDKIM_to_dkimSigResultV2(arhDKIM) {
 	/** @type {dkimSigResultV2} */
-	const dkimSigResult = {};
-	dkimSigResult.version = "2.0";
+	const dkimSigResult = {
+		version: "2.0",
+		result: "PERMFAIL",
+	};
 	switch (arhDKIM.result) {
 		case "none": {
 			dkimSigResult.result = "none";
@@ -140,7 +142,8 @@ function arhDKIM_to_dkimSigResultV2(arhDKIM) {
 			break;
 		}
 		default: {
-			throw new Error(`invalid dkim result in arh: ${arhDKIM.result}`);
+			dkimSigResult.result = "PERMFAIL";
+			dkimSigResult.errorType = arhDKIM.reason ?? arhDKIM.result;
 		}
 	}