+++ test/helpers/chaiUtils.mjs.js
@@ -8,8 +8,9 @@
  */
 
 // @ts-check
+/* eslint-disable jsdoc/reject-any-type */
 
-/** @type {Chai.ExpectStatic} */
+/** @type {import("chai").expect} */
 // @ts-expect-error
 const expect = globalThis.expect;
 export default expect;
@@ -18,7 +19,7 @@ import { DKIM_SigError, DKIM_TempError } from "../../modules/error.mjs.js";
 import Logging from "../../modules/logging.mjs.js";
 
 // disable logging in tests
-Logging.setLogLevel(Logging.Level.Fatal);
+Logging.setLogLevel(Logging.Level.Off);
 
 /**
  * Assert that the given promise is rejected with a certain type of DKIM_SigError.